+++ java/amapvox.lidar.gridded/src/main/java/org/amapvox/lidar/faro/XYBScan.java
@@ -24,6 +24,11 @@ public class XYBScan extends GriddedPointScan {
     }
     
     @Override
+    public void open() throws IOException {
+        readHeader();
+    }
+    
+    @Override
     public void readHeader() throws FileNotFoundException, IOException {
         
         XYBIterator xybIterator = new XYBIterator(getFile());